Move dandi helper functions to pynwb.utils
