+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - Updated CI and tests to handle deprecations in libraries. @rly (#1377)
 - Added test utilities for icephys (``pynwb.testing.icephys_testutils``) to ease creation of test data
   for tests and tutorials. @oruebel (#1349, #1383)
+- Added utility functions for accessing DANDI Archive assets in ``pynwb.utils``. @rly (#1393)
 - Added tests of streaming using the ROS3 driver. @rly (#1393)
 
 ### Bug fixes:
