+++ setup.cfg
@@ -31,6 +31,7 @@ per-file-ignores =
   src/pynwb/legacy/io/__init__.py:F401
   tests/integration/__init__.py:F401
   src/pynwb/testing/__init__.py:F401
+  src/pynwb/utils/__init__.py:F401
   src/pynwb/validate.py:T001
   setup.py:T001
   test.py:T001
